requirements.txt
agora-token-builder==1.0.0
aiogram==2.25.2
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==21.11.1
Automat==20.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.15.0
channels==3.0.4
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.2.16
django-cors-headers==4.4.0
django-gunicorn==0.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-swagger==0.2.2
drf-yasg==1.21.7
environs==8.0.0
exceptiongroup==1.2.2
fastapi==0.115.4
Flask==3.0.3
fractal==0.3.0
frozenlist==1.4.1
functions==0.7.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib_metadata==8.4.0
incremental==21.3.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.1
magic-filter==1.0.12
MarkupSafe==2.1.5
marshmallow==3.21.3
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
Pillow==10.0.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyOpenSSL==21.0.0
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-whatsapp-bot==1.0.92
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
redis==4.6.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
rubicon-objc==0.4.9
selenium==4.24.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.4.2
starlette==0.41.2
telegram==0.0.1
Telethon==1.36.0
termcolor==2.4.0
tomli==2.0.2
transliterate==1.10.2
trio==0.26.2
trio-websocket==0.11.1
twilio==9.3.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.12.2
tzdata==2021.5
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
whatsapp-api-client-python==0.0.45
whatsapp-chatbot-python==0.9.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.1
zope.interface==5.4.0
